Add more info about the site and jwt in general
Know more: decoding is done entirely client side and a list of
registered claim names and their meaning is shown on the page.

Read less: green text for not-yet expired tokens and red text for
expired tokens.

Easier on the eyes: visited links are now blue to add more uniformity to the way the
page looks and the text is never wider than 80 characters.

Threw a jab at jwt.io in there for good measure.