Add highlighting and darkmode
Due to popular demand the decoded jwt payload is now syntax highlighted
with help from https://highlightjs.org.

A new dark mode will kick in by sunset to be easier on the retinas out
there. It works by getting the client's location through
http://ipinfo.io/json, then the lat- and longitude is fed to `suncalc`
(https://github.com/mourner/suncalc) to get the time for when the sunset
begins at the client's location. If the sun has begun to set, the
darkmode is activated.

It's also possible to always use the darkmode by checking the "Prefer
darkmode" box in the bottom of the screen. The preference is saved in
the browser's `localStorage`.

Clear your caches ppl.