Add a changelog
This commit adds a script that will generate a changelog based on the
commit history on each deploy. The output of the script is run through
the markdown parser and the output of that is  used to replace a
`{CHANGELOG}` placeholder in a `template.html` file.

The html is styled with a markdown stylesheet heavily inspired (read
copied, then modified) from https://github.com/mrcoles/markdown-css so
it looks like markdown again. Lastly, all links are found and made into
anchor tags with linkify because I don't want to write markdown-style
links in commit messages (https://github.com/SoapBox/linkifyjs/).

Crazy stuff.