Use class based way of changing color scheme
Previously we disabled a stylesheet to enable darkmode. That required
stylesheets to be included in a special order in `index.html`. Ugh.

This patch gives `<body>` either a `light` or `dark` value for the class
attribute instead. Removes the need for special inclusion order and
makes the code to switch simpler. Yay \o/