Replace query string sharing with hash and add copy
Previously it was possible to share jwts with a value for a `jwt` query
string parameter. This is no longer possible because it showed up in
nginx logs and I'm not interested in anything like that. The hash part
of a url is not logged which is a better option for jwt.show.

We also add a "Copy share link to clipboard" and create a bit of room
between the buttons and the `#encoded` field.