Mithril edition of jwt.show
The initial version of jwt.show was thrown together very quickly and uses some
older techniques for manipulating the DOM in order to achieve showing a decoded
jwt, counting down, copying to clipboard, etc.

It certainly works, but I wondered what was on the other side of the
fence, both in ES6 and virtual DOM land. This commit takes the first
version of jwt.show and implements it in ES8 (async/await) with
[mithril][1] as the framework.

There's undoubtedly still room for improvement but it feels much nicer
than the old code base and I hope it will be easier to build new
features on top of it in the future :-)

[1]: http://mithril.js.org