Housecleaning: move/delete a couple of files
The src folder now contains all the site files. The tools folder now
holds all the tools. I don't really use ctags for this project, so the
tags file is removed. So is a test file that I used for expirements with
generating the changelog with escaped newlines and quotes.