Stop using the deprecated escape function
Previously we used the `escape` function to make sure that special
characters in the jwt was decoded correctly. The function is deprecated
and now we use the alternative suggested by MDN. We also make the
decoded jwt a property rather than a function in `JwtState` to avoid the
function being constantly evaluated.