Escape double quotes in commit subject
Since the changelog is generated json with the git commit log as its
basis it's necessary to escape double quotes. Recently the first commit
message with double quotes was added which caused build time issues.
This commit makes sure that we escape double quotes in the subject as
well as in the body.