Make the site more mobile friendly
It's not often that we need to decode jwts on mobile devices, but the
site might as well look good on there nonetheless. This commit tries to
achieve that by setting the width of the viewport that the site is
rendered in ([Read more][1]) and by making small adjustments to line
lengths, how certain lines break, and to margins.

[1]: https://developer.mozilla.org/en-US/docs/Mozilla/Mobile/Viewport_meta_tag