Add dockerfile
Prepare jwt-show to be served in a docker container on an environment
where nginx-proxy and the letsencrypt companion are running. A default
configuration for nginx is added with the purpose of hitting the
mithril router in jwt-show if the requested uri does not match a file on
the server.

We also regenerate yarn.lock and add license to package.json since
Github found a vulnerability in yarn.lock regarding jquery. Here's to
hoping it's fixed in the new yarn.lock.