Make jwt.show work a little better in firefox
Darkmode now works as intended in firefox who would like us to be
explicit about the body's height (100%).

Copying jwt/links in firefox also works a bit better due to some
tweaking wrt. animation timings.

Pasting a jwt in firefox also shows the beginning of the input field
instead of the end so you'll be relieved when you see that classic
`eyJ0...` and hum along to
[this classic jwt.show tribute by red hot chili peppers](https://www.youtube.com/watch?v=KvlqclaxyP4&feature=youtu.be&t=37).