Rewrite
Rewritten with vue3 and nuxt3 mostly to see if the grass is greener on
the other side --- it's nice enough to stay.

For now it seems most convenient and cheaper to do the hosting on
netlify or something like that instead of the more custom kubernetes
setup I have right now with linode so all dockerfiles, kubernetes
manifests and whatever else was needed for that setup are also deleted.